Sequences

+++ src/irvine/oeis/a203/A203264.java
@@ -53,7 +53,7 @@ public class A203264 extends AbstractSequence {
       mULT = new UpperLeftTriangle(1, 1, -1) {
         @Override
         public Z matrixElement(final int i, final int j) {
-          return mSeq.a((i - mRow0) * (i - mRow0 + 1) / 2 + j - mCol0);
+          return mSeq.a((long) (i - mRow0) * (i - mRow0 + 1) / 2 + j - mCol0);
         }
       };
     }
